cleanup according to cr
